test(parity): split intake and hardening proof
Why:
- intake workflow parity and service hardening proof were still concentrated in two
  oversized files, which made regressions harder to isolate
- the repo standard calls for narrow, bounded tests so parity evidence remains
  reviewable at the behavior level

What:
- replace the broad intake workflow test module with separate apply, routing,
  and package-review workflow files
- replace the mixed hardening module with focused profile and normalization
  service hardening tests plus a narrow shared support seam
- keep the existing parity scenarios intact while moving them into smaller
  failure surfaces

Checks:
- uv run python -m tools.run_quality_gates --full-tests
