+++ tests/unit/application/services/test_intake_plan_routing_workflow.py
@@ -1,0 +1,194 @@
+from __future__ import annotations
+
+from pathlib import Path
+from zipfile import ZIP_DEFLATED, ZipFile
+
+from crypto_reconciliation.application.dtos import IntakePlanRequest
+from crypto_reconciliation.application.services.intake import SourceIntakeService
+from crypto_reconciliation.infrastructure.serialization.filesystem import FilesystemArtifactStore
+
+
+def test_source_intake_service_plans_archive_members_without_copying_them(tmp_path: Path) -> None:
+    incoming_dir = tmp_path / "incoming"
+    incoming_dir.mkdir()
+    archive_path = incoming_dir / "bundle.zip"
+    with ZipFile(archive_path, "w", compression=ZIP_DEFLATED) as archive:
+        archive.writestr("inner.csv", "a,b\n1,2\n")
+
+    workspace_root = tmp_path / "workspace"
+    report_dir = tmp_path / "reports"
+
+    response = SourceIntakeService(FilesystemArtifactStore()).plan(
+        IntakePlanRequest(
+            incoming_dir=incoming_dir,
+            workspace_root=workspace_root,
+            report_dir=report_dir,
+        )
+    )
+
+    plan_rows = FilesystemArtifactStore().read_rows(report_dir / "intake_plan.csv")
+
+    assert response.file_count == 2
+    assert any(row["action"] == "copy" and row["path"].endswith("bundle.zip") for row in plan_rows)
+    assert any(row["action"] == "extract_copy" and row["archive_member_path"] == "inner.csv" for row in plan_rows)
+
+
+def test_source_intake_service_routes_source_artifacts_to_source_aware_supporting_paths(
+    tmp_path: Path,
+) -> None:
+    incoming_dir = tmp_path / "incoming"
+    image_path = incoming_dir / "2021" / "Binance" / "From Binance" / "Trade Analysis - ADA-USDT - Binance.png"
+    scratch_csv = incoming_dir / "2021" / "Binance" / "2021 Isolated" / "test.csv"
+    image_path.parent.mkdir(parents=True, exist_ok=True)
+    scratch_csv.parent.mkdir(parents=True, exist_ok=True)
+    image_path.write_bytes(b"png")
+    scratch_csv.write_text("a,b\n1,2\n", encoding="utf-8")
+
+    workspace_root = tmp_path / "workspace"
+    report_dir = tmp_path / "reports"
+
+    SourceIntakeService(FilesystemArtifactStore()).plan(
+        IntakePlanRequest(
+            incoming_dir=incoming_dir,
+            workspace_root=workspace_root,
+            report_dir=report_dir,
+        )
+    )
+
+    plan_rows = FilesystemArtifactStore().read_rows(report_dir / "intake_plan.csv")
+    by_name = {Path(row["path"]).name: row for row in plan_rows if not row["archive_member_path"]}
+
+    assert by_name["Trade Analysis - ADA-USDT - Binance.png"]["role"] == "working_derivative"
+    assert by_name["Trade Analysis - ADA-USDT - Binance.png"]["source_folder"] == "binance"
+    assert "/working/supporting_artifacts/binance/" in by_name["Trade Analysis - ADA-USDT - Binance.png"]["target_path"]
+    assert by_name["test.csv"]["role"] == "working_derivative"
+    assert "/working/supporting_artifacts/binance/" in by_name["test.csv"]["target_path"]
+
+
+def test_source_intake_service_routes_cointracking_html_and_sidecar_to_portfolio_capture(
+    tmp_path: Path,
+) -> None:
+    incoming_dir = tmp_path / "incoming"
+    html_path = incoming_dir / "tmp" / "CoinTracking · Tax Declaration Export.html"
+    sidecar_path = incoming_dir / "tmp" / "CoinTracking · Tax Declaration Export_files" / "style.min.css"
+    sidecar_path.parent.mkdir(parents=True, exist_ok=True)
+    html_path.write_text(
+        """
+        <html>
+        <head><title>CoinTracking · Tax Declaration Export</title></head>
+        <body>Created by: CoinTracking as of: 06.04.2022 01:11</body>
+        </html>
+        """,
+        encoding="utf-8",
+    )
+    sidecar_path.write_text("body{}", encoding="utf-8")
+
+    workspace_root = tmp_path / "workspace"
+    report_dir = tmp_path / "reports"
+
+    SourceIntakeService(FilesystemArtifactStore()).plan(
+        IntakePlanRequest(
+            incoming_dir=incoming_dir,
+            workspace_root=workspace_root,
+            report_dir=report_dir,
+        )
+    )
+
+    plan_rows = FilesystemArtifactStore().read_rows(report_dir / "intake_plan.csv")
+    by_name = {Path(row["path"]).name: row for row in plan_rows}
+
+    assert by_name["CoinTracking · Tax Declaration Export.html"]["role"] == "portfolio_export"
+    assert by_name["CoinTracking · Tax Declaration Export.html"]["capture_id"] == "2022-04"
+    assert (
+        "/evidence/raw/portfolio/cointracking/2022-04/"
+        in by_name["CoinTracking · Tax Declaration Export.html"]["target_path"]
+    )
+    assert by_name["style.min.css"]["role"] == "portfolio_sidecar"
+    assert by_name["style.min.css"]["capture_id"] == "2022-04"
+
+
+def test_source_intake_service_routes_wallet_export_to_existing_inventory_source(tmp_path: Path) -> None:
+    workspace_root = tmp_path / "workspace"
+    source_inventory_path = workspace_root / "analysis" / "issues" / "source_inventory.csv"
+    source_inventory_path.parent.mkdir(parents=True, exist_ok=True)
+    FilesystemArtifactStore().write_rows(
+        source_inventory_path,
+        (
+            "source",
+            "status",
+            "capture_path",
+            "adapter",
+            "normalization_status",
+        ),
+        (
+            {
+                "source": "eth-gala1",
+                "status": "capture_complete",
+                "capture_path": "evidence/raw/source/eth-gala1/2026-03",
+                "adapter": "evm_explorer",
+                "normalization_status": "ready",
+            },
+        ),
+    )
+    wallet_evidence_path = workspace_root / "analysis" / "inventory" / "wallet_inventory_evidence.csv"
+    wallet_evidence_path.parent.mkdir(parents=True, exist_ok=True)
+    FilesystemArtifactStore().write_rows(
+        wallet_evidence_path,
+        (
+            "source",
+            "capture_path",
+            "wallet_id",
+            "identifier_kind",
+            "normalized_identifier",
+            "display_identifier",
+            "network_scope",
+            "controller",
+            "account_label",
+            "evidence_kind",
+            "evidence_path",
+            "confidence",
+            "note",
+        ),
+        (
+            {
+                "source": "eth-gala1",
+                "capture_path": "/tmp/capture",
+                "wallet_id": "evm_address:0x2222222222222222222222222222222222222222",
+                "identifier_kind": "evm_address",
+                "normalized_identifier": "0x2222222222222222222222222222222222222222",
+                "display_identifier": "0x2222222222222222222222222222222222222222",
+                "network_scope": "ethereum",
+                "controller": "Explorer export",
+                "account_label": "Account 2",
+                "evidence_kind": "filename",
+                "evidence_path": "/tmp/evidence.csv",
+                "confidence": "high",
+                "note": "",
+            },
+        ),
+    )
+
+    incoming_dir = tmp_path / "incoming"
+    incoming_dir.mkdir()
+    export_path = incoming_dir / "Account1-bsc export-address-token.csv"
+    export_path.write_text(
+        "Transaction Hash,Blockno,UnixTimestamp,DateTime (UTC),TokenValue,TokenSymbol,From,To\n"
+        "0xabc,1,1710000000,2024-03-09 09:41:37,1,GALA,0x0,0x2222222222222222222222222222222222222222\n",
+        encoding="utf-8",
+    )
+    report_dir = tmp_path / "reports"
+
+    SourceIntakeService(FilesystemArtifactStore()).plan(
+        IntakePlanRequest(
+            incoming_dir=incoming_dir,
+            workspace_root=workspace_root,
+            report_dir=report_dir,
+        )
+    )
+
+    plan_rows = FilesystemArtifactStore().read_rows(report_dir / "intake_plan.csv")
+    row = next(item for item in plan_rows if item["archive_member_path"] == "")
+
+    assert row["source_folder"] == "eth-gala1"
+    assert row["inventory_match_status"] == "inventory_source_match"
+    assert row["review_required"] == "no"
